requirements.txt
requests
qrcode
Pillow